Suppress the warnings reported by cppcheck
In C99/C11 6.5.7/4. In the result of E1 << E2, if E1 has a signed type
and nonnegative value, and E1 * 2 to the power of E2 is representable
in the result type, then that is the resulting value; otherwise, the
behavior is undefined. For example:

 - "1 << 31" is undefined because 2147483648 can't be represented in
    a 32 bit signed integer.

 - "1u << 31" is well defined because 2147483648 can be represented in
    a 32 bit unsigned integer.

In C99/C11 6.5.7/5. In the result of E1 >> E2, if E1 has a signed type
and a negative value, the resulting value is implementation-defined.
For example:

 - "-1 >> 31" is "-1" (performs right shift arithmetically).

 - "-1 >> 31" is "1"  (performs right shift logically).

In datalab the rule say that we can assume machine performs right shift
arithmetically. However, The cppcheck will complain when shifting signed
32-bit value by 31 bits, this commit change these variables that needs
to shift 31 bits to become unsigned and add the integer-suffix 'u' to
integer constants. For example:

 - "x << 31" becomes "(unsigned) x << 31"

 - "x >> 31" becomes "(unsigned) x >> 31"

 - "1 << 31" becomes "1u << 31"